requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=23.3,<25.0
Pillow>=10.4,<11.0
psycopg2-binary>=2.9,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
django-cors-headers>=4.4,<5.0
